requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.9
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
